Deleting tests folder -- all the checks are already done
